==> src/02a_nf-resolveome_setup.R <==
#!/usr/bin/env Rscript

# libraries
library(magrittr)

# dirs
donor_id_i <- "PD63118"
wd <- getwd()
out_dir <- file.path(wd, "out/nf-resolveome/")
dir.create(out_dir, showWarnings = FALSE)

# get clean cells only (and cells that have not yet been assessed)
clean_cell_ids <-
  readr::read_tsv("data/manual_inspection/2024-12-20_PD63118_PTA_BAF_LoH_CellType_Mut_Summary.tsv") %>%
  dplyr::filter(!suspected_doublet | is.na(suspected_doublet),
                !chr_dropout | is.na(chr_dropout)) %>%
  dplyr::pull(cell_id)

# samplesheet
ss_bams <-
  readr::read_csv("data/resolveome/samplesheet_local.csv") %>%
  dplyr::filter(seq_type %in% c("dna", "dnahyb"), cell_id %in% clean_cell_ids)

# function: define mutation type based on ref and alt, split up mnvs and dnvs
type_mutations <- function(df) {
  typed_df <-
    df %>%
    dplyr::mutate(type = dplyr::case_when(
      nchar(ref) == 1 & nchar(alt) == 1 ~ "snv",
      nchar(ref) == 1 & nchar(alt) > 1 ~ "ins",
      nchar(ref) > 1 & nchar(alt) == 1 ~ "del",
      nchar(ref) == 2 & nchar(alt) == 2 ~ "dnv",
      nchar(ref) > 2 & nchar(alt) > 2 ~ "mnv"
    ))

  # expand dnv/mnv mutations to all positions
  if (any(typed_df$type %in% c("dnv", "mnv"))) {
    typed_df_dnv_mnv <-
      typed_df %>%
      dplyr::filter(type %in% c("dnv", "mnv")) %>%
      dplyr::mutate(mut_id = paste(chr, pos, ref, alt, type)) %>%
      dplyr::group_by(dplyr::across(-c(pos, alt, ref))) %>%
      dplyr::reframe(pos = pos:(pos + nchar(alt) - 1),
                     ref = strsplit(ref, split = "") %>% unlist(),
                     alt = strsplit(alt, split = "") %>% unlist()) %>%
      dplyr::select(-mut_id)
    typed_df <-
      typed_df %>%
      dplyr::filter(!(type %in% c("dnv", "mnv"))) %>%
      dplyr::bind_rows(typed_df_dnv_mnv)
  }

  # return
  typed_df
}

# get common snp sites
common_snps <-
  gzfile("/lustre/scratch125/casm/team268im/fa8/117/NOVASEQX_MASKS/NEW_MASKS_wNSX_OCT2024/GRCh37_WGNS/SNP_GRCh37.wgns.bed.gz") %>%
  readr::read_tsv(col_names = c("#CHROM", "START", "POS")) %>%
  dplyr::select(`#CHROM`, POS)

# get a caveman snp file from a sample with high coverage
# extract common snps that are heterozygous
# 0.3 < VAF < 0.7 and DP > 50
# PD63118b_lo0044 has the highest coverage at 68X according to picard
caveman_snps <-
  "/nfs/cancer_ref01/nst_links/live/3464/PD63118b_lo0044/PD63118b_lo0044.caveman_c.snps.vcf.gz" %>%
  readr::read_tsv(comment = "##") %>%
  dplyr::mutate(
    DP = strsplit(INFO, ";") %>% purrr::map_chr(~ .x[grepl("^DP=", .x)]) %>%
      strsplit("=") %>% purrr::map_chr(~ .x[2]) %>% as.integer(),
    VAF = gsub(".*:", "", TUMOUR) %>% as.numeric()) %>%
  dplyr::filter(DP > 50, 0.3 < VAF, VAF < 0.7) %>%
  # get those at common snp sites
  dplyr::inner_join(common_snps) %>%
  dplyr::transmute(donor_id = donor_id_i, chr = `#CHROM`, pos = POS, ref = REF,
                   alt = ALT) %>%
  # type the mutations
  type_mutations() %>%
  dplyr::distinct()

# get mutations
nanoseq_muts <-
  readr::read_csv("data/nanoseq/mutations2genotype_20250219.csv") %>%
  # get unique mutations
  dplyr::distinct() %>%
  dplyr::mutate(donor_id = donor_id_i) %>%
  # type the mutations
  type_mutations() %>%
  # remove duplicates
  dplyr::distinct(chr, pos, ref, alt, donor_id)

# write mutations
snps_file <- file.path(out_dir, donor_id_i, "caveman_snps.tsv")
readr::write_tsv(caveman_snps, snps_file)
muts_file <- file.path(out_dir, donor_id_i, "nanoseq_mutations.tsv")
readr::write_tsv(nanoseq_muts, muts_file)

# write samplesheets
ss <-
  ss_bams %>%
  dplyr::mutate(mutations = muts_file, snps = snps_file) %>%
  {split(., .$seq_type)}
purrr::walk2(names(ss), ss, function(seq_type_i, ss_i) {
  ss_i %>%
    readr::write_csv(file.path(out_dir, seq_type_i, "samplesheet.csv"))
})
